# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminLogin": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "AdminLogin",
        "type": "object"
      },
...
      "EmailVerificationRequest": {
        "additionalProperties": false,
        "description": "Present a verification link's token.\n\nNo account identifier: possession of the token IS the claim, and accepting\na user id alongside it would create a second, weaker way to say which\naccount is being verified.",
        "properties": {
          "token": {
            "maxLength": 512,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "EmailVerificationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PasswordResetCompletion": {
        "additionalProperties": false,
        "description": "Present a link's token and choose a new password.",
        "properties": {
          "new_password": {
            "maxLength": 200,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 512,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordResetCompletion",
        "type": "object"
      },
      "PasswordResetRequest": {
        "additionalProperties": false,
        "description": "Ask for a reset link. Answered identically whatever the address is.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequest",
        "type": "object"
      },
...
      "RecoveryAccepted": {
        "description": "The one answer every message-sending recovery endpoint gives.\n\nA FIXED STRING, not a rendered outcome. The moment this carries a field\nthat varies \u2014 \"sent\": true, an address, a masked address, a count \u2014 the\nendpoint starts answering \"does this address have an account?\", which is\nthe single thing these routes exist not to answer.",
        "properties": {
          "detail": {
            "default": "If that address has an account, a message is on its way. Check your inbox, including spam.",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "RecoveryAccepted",
        "type": "object"
      },
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationResult": {
        "description": "What a completed verification tells the client.\n\n`status` so the SPA can drop its \"unverified\" banner without a second\nround trip. It is the account's own status, which the caller just changed\nand is entitled to see.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "VerificationResult",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered Canadian tax optimization SaaS \u2014 educational, not filing.",
    "title": "Onyx Ledger API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/auth/login": {
      "post": {
        "description": "Operator login.\n\nThrottled on the same class as user login, and if anything it matters more\nhere: the population of valid operator addresses is tiny, which makes this\nthe highest-value guessing target in the product.",
        "operationId": "admin_login_api_v1_admin_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Login Api V1 Admin Auth Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Login",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Exchange credentials for a token pair.\n\nAdmission runs FIRST \u2014 before the user lookup and before Argon2 \u2014 so a\nrefused attempt costs one indexed UPSERT instead of a deliberately expensive\nkey derivation. See `app.services.admission.auth`.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password-reset": {
      "post": {
        "description": "Ask for a reset link.\n\nNON-ENUMERATING, and the handler is written so that it cannot accidentally\nstop being so. There is one `return` and it is unconditional; the service\nanswers `None` for every address it will not mail \u2014 unknown, suspended,\nclosed, deleting, or asked for too recently \u2014 and `deliver(None)` does\nnothing. No branch in this function can observe which of those happened, so\nno future edit can leak it by adding a field to one arm.\n\n202 rather than 200: the honest status for \"your request is accepted and\nsomething may happen later\", and identical either way.",
        "operationId": "request_password_reset_api_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Password Reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "description": "Set a new password from a reset link, and end every existing session.\n\nAnonymous for the same reason as verification: the link is clicked wherever\nthe mail was opened, and the token is the claim.\n\n204 and no body. Returning a token pair here would be convenient and wrong\n\u2014 it would hand a session to whoever holds the link, in the one flow whose\npremise is that somebody else may have had access. The customer signs in\nwith the password they just chose, which also proves they know it.",
        "operationId": "complete_password_reset_api_v1_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetCompletion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Password Reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate a refresh token.\n\nAlso a credential surface: a stolen or guessed refresh token is presented\nhere. The identity scope is keyed on the presented TOKEN rather than on an\nemail, because that is the only identity claim the request carries \u2014 it\nbounds how fast one token can be hammered, while the address scope bounds\nhammering with a rotating supply of them.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Create an account and send it a verification link.\n\nThe account is created unverified: it can sign in, and it cannot reach the\nrest of the API until the emailed link is opened.\n\nThrottled on the same class as login. Registration is not a credential test,\nbut it is unauthenticated, it writes two rows and computes an Argon2 hash,\nand \u2014 because it must say whether an address is already taken \u2014 an\nunthrottled version is a fast account-enumeration oracle. The throttle does\nnot remove that disclosure; it bounds how quickly it can be harvested.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Api V1 Auth Register Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verification": {
      "post": {
        "description": "Send, or re-send, this account's verification link.\n\nAUTHENTICATED, and that is what makes it safe to offer at all. An anonymous\n\"send verification to this address\" endpoint is an enumeration oracle and an\nemail-flood amplifier pointed at anybody whose address you can guess; this\none can only ever mail the account whose token was presented.\n\n`db_authed_unverified_ok` rather than `db_authed`: this is the endpoint that\nclears the unverified state, so it is the one that cannot require it to be\nclear already. Suspension, closure and the deletion cutoff all still apply.\n\nThrottled on `ACCOUNT_RECOVERY` keyed on the ACCOUNT, not the address \u2014 the\naddress would have to be read from the user table before the limiter ran,\nand they are one to one anyway.\n\n202, and the same body whatever happened. An account that is already\nverified mints nothing and sends nothing, and says so no differently.",
        "operationId": "send_verification_api_v1_auth_verification_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryAccepted"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Verification",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verification/confirm": {
      "post": {
        "description": "Complete verification by presenting a link's token.\n\nANONYMOUS on purpose: the link is clicked in whatever browser opened the\nmail, which is routinely not the one holding a session. Possession of the\ntoken is the entire claim, and requiring a bearer token as well would break\nthe common case to add nothing \u2014 an attacker holding the token does not\nalso need to be signed in.\n\nThrottled on `AUTH_ATTEMPT` keyed on the presented token, exactly as\n`/refresh` is and for the same reason: this is a credential surface, and the\ncredential is the only identity claim the request carries.",
        "operationId": "confirm_verification_api_v1_auth_verification_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Verification",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}